worked on transforming 3d Coordinates into pixel coordinates

+++ src/screen_calculations.c
@@ -6,7 +6,7 @@
 /*   By: afoth <afoth@student.42berlin.de>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 16:00:16 by afoth             #+#    #+#             */
-/*   Updated: 2025/01/21 17:27:06 by afoth            ###   ########.fr       */
+/*   Updated: 2025/02/24 18:44:39 by afoth            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 void calc_aspect_ratio(t_rt *rt)
 {
 	if(SCREEN_HEIGHT > 0 && SCREEN_WIDTH > 0)
-		rt->aspect_r = SCREEN_HEIGHT / SCREEN_WIDTH;
+		rt->aspect_r = (float)SCREEN_HEIGHT / (float)SCREEN_WIDTH;
 	else
 		ft_exit(rt, 4, "Screen size zero or negative");
 }
